# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "API Absensi Siswa",
    "description": "Dokumentasi API untuk sistem absensi RFID siswa",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "bearerAuth": {
      "type": "http",
      "scheme": "bearer",
      "bearerFormat": "JWT"
    }
  },
  "paths": {
...
    "/api/absen-siswa": {
      "post": {
        "tags": [
          "Absensi"
        ],
        "description": "Absen masuk/pulang siswa berdasarkan kartu RFID",
        "responses": {
          "200": {
            "description": "Berhasil absen"
          },
          "400": {
            "description": "Validasi gagal atau sudah absen"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Gagal proses absen"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "card_uid": {
                    "type": "string"
                  }
                },
                "required": [
                  "card_uid"
                ]
              }
            }
          }
        }
      }
    },
    "/api/generate-absen-harian": {
      "post": {
        "tags": [
          "Absensi"
        ],
        "description": "Generate absen harian default dengan status alpha",
        "responses": {
          "201": {
            "description": "Data absen berhasil dibuat"
          },
          "500": {
            "description": "Gagal generate absen"
          }
        }
      }
    },
    "/api/setting": {
      "get": {
        "tags": [
          "Setting"
        ],
        "description": "Ambil data pengaturan absen",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Berhasil ambil pengaturan"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/edit-setting": {
      "put": {
        "tags": [
          "Setting"
        ],
        "description": "Edit pengaturan waktu absen",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Pengaturan berhasil diperbarui"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Pengaturan tidak ditemukan"
          },
          "500": {
            "description": "Gagal edit pengaturan"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status_manual": {
                    "type": "boolean"
                  },
                  "status_otomatis": {
                    "type": "boolean"
                  },
                  "jam_masuk": {
                    "type": "string"
                  },
                  "jam_keluar": {
                    "type": "string"
                  },
                  "batas_absen_masuk": {
                    "type": "string"
                  },
                  "batas_absen_keluar": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}